add styles for a button
